requirements.txt
asgiref==3.4.1
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.9
django-filter==21.1
django-guardian==2.4.0
django-softdelete==0.9.2
djangorestframework==3.12.4
docopt==0.6.2
idna==3.3
iniconfig==1.1.1
Markdown==3.3.4
model-bakery==1.3.3
packaging==21.2
pathlib==1.0.1
pluggy==1.0.0
psycopg2==2.9.1
py==1.11.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
yarg==0.1.9
